requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
